@@ -75,63 +77,60 @@
     "reflect-metadata": "^0.1.13",
     "ts-loader": "^8.3.0",
     "ts-node": "^9.1.1",
-    "tsyringe": "^4.5.0",
-    "typescript": "^4.3.4",
+    "tsyringe": "^4.6.0",
+    "typescript": "^4.3.5",
     "vue-apollo": "^3.0.7",
-    "vue-tailwind": "^2.2.1",
+    "vue-tailwind": "^2.4.1",
     "vue-tree-list": "^1.5.0"
   },
   "devDependencies": {
     "@babel/core": "^7.14.6",
-    "@graphql-codegen/cli": "1.21.5",
-    "@graphql-codegen/typescript": "^1.22.2",
-    "@graphql-codegen/typescript-operations": "^1.18.2",
-    "@graphql-codegen/typescript-resolvers": "^1.19.4",
-    "@graphql-codegen/typescript-vue-apollo": "^2.3.7",
+    "@graphql-codegen/cli": "^1.21.7",
+    "@graphql-codegen/fragment-matcher": "^2.0.1",
+    "@graphql-codegen/typed-document-node": "^1.18.10",
+    "@graphql-codegen/typescript": "^1.22.4",
+    "@graphql-codegen/typescript-operations": "^1.18.4",
+    "@graphql-codegen/typescript-resolvers": "^1.20.0",
+    "@graphql-eslint/eslint-plugin": "^1.1.3",
     "@graphql-inspector/cli": "^2.6.1",
+    "@graphql-typed-document-node/core": "^3.1.0",
     "@nuxt/postcss8": "^1.1.3",
     "@nuxt/types": "^2.15.7",
     "@nuxt/typescript-build": "^2.0.6",
     "@nuxtjs/eslint-config": "^6.0.1",
     "@nuxtjs/eslint-config-typescript": "^6.0.1",
     "@nuxtjs/eslint-module": "^3.0.2",
-    "@nuxtjs/tailwindcss": "4.2.0",
+    "@nuxtjs/tailwindcss": "4.2.1",
     "@tailwindcss/forms": "^0.2.1",
     "@tailwindcss/line-clamp": "^0.2.0",
     "@types/bcryptjs": "^2.4.2",
-    "@types/connect-redis": "^0.0.16",
-    "@types/express": "^4.17.12",
-    "@types/express-session": "^1.17.3",
-    "@types/jest": "^26.0.23",
-    "@types/lodash": "^4.14.169",
-    "@types/nodemailer": "^6.4.2",
-    "@types/passport": "^1.0.6",
-    "@types/uuid": "^8.3.0",
-    "@types/yaireo__tagify": "^4.3.0",
-    "@vue/compiler-sfc": "^3.1.2",
+    "@types/connect-redis": "^0.0.17",
+    "@types/express": "^4.17.13",
+    "@types/express-session": "^1.17.4",
+    "@types/jest": "^26.0.24",
+    "@types/lodash": "^4.14.171",
+    "@types/nodemailer": "^6.4.4",
+    "@types/passport": "^1.0.7",
+    "@types/uuid": "^8.3.1",
+    "@types/yaireo__tagify": "^4.3.2",
+    "@vue/compiler-sfc": "^3.1.5",
     "@vue/test-utils": "^1.2.1",
     "babel-jest": "^26.6.3",
     "concurrently": "^6.2.0",
     "cross-env": "^7.0.3",
-    "eslint": "^7.29.0",
+    "eslint": "^7.30.0",
     "eslint-config-prettier": "^8.3.0",
     "eslint-plugin-nuxt": "^2.0.0",
     "eslint-plugin-prettier": "^3.4.0",
-    "eslint-plugin-unused-imports": "^1.1.1",
-    "eslint-plugin-vue": "^7.12.1",
+    "eslint-plugin-unused-imports": "^1.1.2",
+    "eslint-plugin-vue": "^7.13.0",
     "jest": "^26.6.3",
-    "jest-mock-extended": "^1.0.16",
+    "jest-mock-extended": "^1.0.18",
     "nuxt-vite": "^0.1.1",
     "prettier": "^2.3.2",
-    "prisma": "2.25.0",
+    "prisma": "2.27.0",
     "ts-jest": "^26.5.6",
     "vue-jest": "4.0.1"
-  },
-  "resolutions": {
-    "ts-invariant": "0.6.0"
-  },
-  "resolutionsComment": {
-    "ts-invariant": "ts-invariant 0.4.0 is required by apollo-link, which is required by graphql-tools v4, which is required by apollo-server. In v3 of apollo-server, the dependency of graphql-tools is finally updated, and this workaround can be removed."
   },
   "workspaces": {
     "nohoist": [
