@@ -1,24 +1,19 @@
-import { container } from 'tsyringe'
+import { mergeResolvers } from '@graphql-tools/merge'
 import { DateTimeResolver, EmailAddressResolver } from 'graphql-scalars'
-import _ from 'lodash'
 import { Resolvers } from './graphql'
-import { Resolvers as UserResolvers } from './user/resolvers'
-import { Resolvers as DocumentResolvers } from './documents/resolvers'
-import { Resolvers as GroupResolvers } from './groups/resolvers'
+import { resolvers as userResolvers } from './user/resolvers'
+import { resolvers as documentResolvers } from './documents/resolvers'
+import { resolvers as groupResolvers } from './groups/resolvers'
 
-const userResolvers = container.resolve(UserResolvers)
-const documentResolvers = container.resolve(DocumentResolvers)
-const groupResolvers = container.resolve(GroupResolvers)
-
-const resolvers: Resolvers = _.merge(
-  userResolvers.resolvers(),
-  documentResolvers.resolvers(),
-  groupResolvers.resolvers(),
-  {
-    // Custom scalar types
-    DateTime: DateTimeResolver,
-    EmailAddress: EmailAddressResolver,
-  }
-)
-
-export default resolvers
+export function loadResolvers(): Resolvers {
+  return mergeResolvers([
+    userResolvers(),
+    documentResolvers(),
+    groupResolvers(),
+    {
+      // Custom scalar types
+      DateTime: DateTimeResolver,
+      EmailAddress: EmailAddressResolver,
+    },
+  ])
+}
