@@ -2,7 +2,7 @@ import { loadSchemaSync as loadGraphqlSchemaSync } from '@graphql-tools/load'
 import { GraphQLFileLoader } from '@graphql-tools/graphql-file-loader'
 import { addResolversToSchema } from '@graphql-tools/schema'
 import { GraphQLSchema } from 'graphql'
-import resolvers from './resolvers'
+import { loadResolvers } from './resolvers'
 
 /**
  * Synchronously loads the schema and the resolvers.
@@ -12,6 +12,14 @@ export function loadSchema(): GraphQLSchema {
   const typeDefs = loadGraphqlSchemaSync('./api/**/*.graphql', {
     loaders: [new GraphQLFileLoader()],
   })
+  const resolvers = loadResolvers()
 
-  return addResolversToSchema(typeDefs, resolvers)
+  return addResolversToSchema({
+    schema: typeDefs,
+    resolvers,
+    resolverValidationOptions: {
+      // Ignore additional methods in our resolver classes (e.g. private fields)
+      requireResolversToMatchSchema: 'ignore',
+    },
+  })
 }
