@@ -1,17 +1,17 @@
+/* eslint-disable no-console */
 import { PrismaClient, GroupType, GroupHierarchyType } from '@prisma/client'
 
 async function seedInternal(prisma: PrismaClient): Promise<void> {
   await prisma.user.deleteMany({})
-  // Manually delete other fields first (at the moment, its not possible to do this automatically https://github.com/prisma/prisma/issues/2810)
-  await prisma.userDocumentOtherField.deleteMany({})
   await prisma.userDocument.deleteMany({})
+  await prisma.userDocumentOtherField.deleteMany({})
   await prisma.group.deleteMany({})
 
   await prisma.user.create({
     data: {
       id: 'ckn4oul7100004cv7y3t94n8j',
-      email: `test@testum.de2`,
-      password: `$2a$10$d.P8dAo.0pmsvW2xwsH7sudRMhSBiCvOfEkN8JBeuvHBs1ZtLUdoe`, // EBNPXY35TYkYXHs
+      email: 'alice@jabref.de',
+      password: '$2a$10$d.P8dAo.0pmsvW2xwsH7sudRMhSBiCvOfEkN8JBeuvHBs1ZtLUdoe', // EBNPXY35TYkYXHs
       name: 'Alice',
     },
   })
@@ -20,12 +20,90 @@ async function seedInternal(prisma: PrismaClient): Promise<void> {
     data: {
       id: 'ckondtcaf000101mh7x9g4gia',
       type: 'Article',
-      title:
-        'Clebsch-Lagrange variational principle and geometric constraint analysis of relativistic field theories',
-      author: 'Tobias Diez and Gerd Rudolph',
+      title: 'Cocoa and Cardiovascular Health',
+      author:
+        'Corti, Roberto and Flammer, Andreas J. and Hollenberg, Norman K. and Luscher, Thomas F.',
       abstract:
-        'Inspired by the Clebsch optimal control problem, we introduce a new variational principle that is suitable for capturing the geometry of relativistic field theories with constraints related to a gauge symmetry. Its special feature is that the Lagrange multipliers couple to the configuration variables via the symmetry group action. The resulting constraints are formulated as a condition on the momentum map of the gauge group action on the phase space of the system. We discuss the Hamiltonian picture and the reduction in the gauge symmetry by stages in this geometric setting. We show that the Yang–Mills–Higgs action and the Einstein–Hilbert action fit into this new framework after a (1 + 3)-splitting. Moreover, we recover the Gaus constraint of Yang–Mills–Higgs theory and the diffeomorphism constraint of general relativity as momentum map constraints.Inspired by the Clebsch optimal control problem, we introduce a new variational principle that is suitable for capturing the geometry of relativistic field theories with constraints related to a gauge symmetry. Its special feature is that the Lagrange multipliers couple to the configuration variables via the symmetry group action. The resulting constraints are formulated as a condition on the momentum map of the gauge group action on the phase space of the system. We discuss the Hamiltonian picture and the reduction in the gauge symmetry by stages in this geometric setting. We show that the Yang–Mills–Higgs action and the Einstein–Hilbert action fit into this new framework after a (1 + 3)-splitting. Moreover, we recover the Gaus constraint of Yang–Mills–Higgs theory and the diffeomorphism constraint of general relativity as momentum map constraints.',
-      keywords: 'Diffeomorphism constraint, gauge symmetry',
+        'Epidemiological data demonstrate that regular dietary intake of plant-derived foods and beverages reduces the risk of coronary heart disease and stroke. Among many ingredients, cocoa might be an important mediator. Indeed, recent research demonstrates a beneficial effect of cocoa on blood pressure, insulin resistance, and vascular and platelet function. Although still debated, a range of potential mechanisms through which cocoa might exert its benefits on cardiovascular health have been proposed, including activation of nitric oxide and antioxidant and antiinflammatory effects. This review summarizes the available data on the cardiovascular effects of cocoa, outlines potential mechanisms involved in the response to cocoa, and highlights the potential clinical implications associated with its consumption. ( Circulation. 2009; 119: 1433-1441.)',
+      keywords: 'cocoa, endothelium, hypertension, platelets',
+      doi: '10.1161/CIRCULATIONAHA.108.827022',
+      volume: '119',
+      issue: '10',
+      pages: '1433-1441',
+      journal: 'Circulation',
+      year: '2009',
+    },
+  })
+  await prisma.userDocument.create({
+    data: {
+      id: 'ckr9eq4oc000101mk1ga9bxnt',
+      type: 'Article',
+      title: 'Cocoa and health: a decade of research',
+      author:
+        'Cooper, Karen A. and Donovan, Jennifer L. and Waterhouse, Andrew L. and Williamson, Gary',
+      abstract:
+        'It has been over 10 years since the first mention in a medical journal about cocoa and chocolate as potential sources of antioxidants for health. During this time, cocoa has been found to improve antioxidant status, reduce inflammation and correlate with reduced heart disease risk; with these results, and its popularity, it has received wide coverage in the press. However, after 10 years of research, what is known about the potential health benefits of cocoa and what are the important next steps in understanding this decadent source of antioxidants?',
+      keywords: 'cocoa, chocolate, health, polyphenols, antioxident',
+      doi: '10.1017/S0007114507795296',
+      volume: '99',
+      issue: '1',
+      pages: '1-11',
+      journal: 'British Journal of Nutrition',
+      year: '2008',
+    },
+  })
+  await prisma.userDocument.create({
+    data: {
+      id: 'ckr9eqap6000301mk20hycjqb',
+      type: 'Article',
+      title:
+        'Chocolate and prevention of cardiovascular disease: A systematic review',
+      author:
+        'Ding, Eric L. and Hutfless, Susan M. and Ding, Xin and Girotra, Saket',
+      abstract: `
+        Background: Consumption of chocolate has been often hypothesized to reduce the risk of cardiovascular disease (CVD) due to chocolate's high levels of stearic acid and antioxidant flavonoids. However, debate still lingers regarding the true long term beneficial cardiovascular effects of chocolate overall.
+        Methods: We reviewed English-language MEDLINE publications from 1966 through January 2005 for experimental, observational, and clinical studies of relations between cocoa, cacao, chocolate, stearic acid, flavonoids ( including flavonols, flavanols, catechins, epicatechins, and procynadins) and the risk of cardiovascular disease ( coronary heart disease (CHD), stroke). A total of 136 publications were selected based on relevance, and quality of design and methods. An updated meta-analysis of flavonoid intake and CHD mortality was also conducted.
+        Results: The body of short-term randomized feeding trials suggests cocoa and chocolate may exert beneficial effects on cardiovascular risk via effects on lowering blood pressure, anti-inflammation, anti-platelet function, higher HDL, decreased LDL oxidation. Additionally, a large body of trials of stearic acid suggests it is indeed cholesterol-neutral. However, epidemiologic studies of serum and dietary stearic acid are inconclusive due to many methodologic limitations. Meanwhile, the large body of prospective studies of flavonoids suggests the flavonoid content of chocolate may reduce risk of cardiovascular mortality. Our updated meta-analysis indicates that intake of flavonoids may lower risk of CHD mortality, RR = 0.81 (95% CI: 0.71 - 0.92) comparing highest and lowest tertiles.
+        Conclusion: Multiple lines of evidence from laboratory experiments and randomized trials suggest stearic acid may be neutral, while flavonoids are likely protective against CHD mortality. The highest priority now is to conduct larger randomized trials to definitively investigate the impact of chocolate consumption on long-term cardiovascular outcomes.`,
+      keywords: 'dark chocolate, flavonoid intake',
+      doi: '10.1186/1743-7075-3-2',
+      volume: '3',
+      issue: '2',
+      journal: 'Nutrition & Metabolism',
+      year: '2006',
+    },
+  })
+  await prisma.userDocument.create({
+    data: {
+      id: 'ckonduhjk000701mh12wia4nf',
+      type: 'Article',
+      title: 'Cocoa and Chocolate in Human Health and Disease',
+      author: 'Katz, David L. and Doughty, Kim and Ali, Ather',
+      abstract:
+        'Cocoa contains more phenolic antioxidants than most foods. Flavonoids, including catechin, epicatechin, and procyanidins predominate in antioxidant activity. The tricyclic structure of the flavonoids determines antioxidant effects that scavenge reactive oxygen species, chelate Fe2+ and Cu+, inhibit enzymes, and upregulate antioxidant defenses. The epicatechin content of cocoa is primarily responsible for its favorable impact on vascular endothelium via its effect on both acute and chronic upregulation of nitric oxide production. Other cardiovascular effects are mediated through anti-inflammatory effects of cocoa polyphenols, and modulated through the activity of NF-kappa B. Antioxidant effects of cocoa may directly influence insulin resistance and, in turn, reduce risk for diabetes. Further, cocoa consumption may stimulate changes in redox-sensitive signaling pathways involved in gene expression and the immune response. Cocoa can protect nerves from injury and inflammation, protect the skin from oxidative damage from UV radiation in topical preparations, and have beneficial effects on satiety, cognitive function, and mood. As cocoa is predominantly consumed as energy-dense chocolate, potential detrimental effects of overconsumption exist, including increased risk of weight gain. Overall, research to date suggests that the benefits of moderate cocoa or dark chocolate consumption likely outweigh the risks.',
+      keywords: 'dark chocolate, blood pressure',
+      doi: '10.1089/ars.2010.3697',
+      volume: '15',
+      issue: '10',
+      pages: '2779-2811',
+      journal: 'Antioxidants & Redox Signaling',
+      year: '2011',
+    },
+  })
+  await prisma.userDocument.create({
+    data: {
+      id: 'ckondu6bh000501mh2o2tf00u',
+      type: 'InProceedings',
+      title: 'Chocolate: food as medicine/medicine as food',
+      author: 'Keen, Carl L.',
+      abstract:
+        'Cocoa and chocolate products have been delicacies for hundreds of years. Only recently have they been recognized as significant sources of phytochemicals with healthful effects. These foods are among the most concentrated sources of the procyanidin flavonoids, catechin and epicatechin. Recent studies have shown that these polyphenols are absorbed from the intestine of animals and humans with epicatechin absorbed much more than catechin. These flavonoids have potent antioxidant and antiplatelet activities following consumption of cocoa or chocolate.',
+      keywords:
+        'chocolate, flavonoids, antioxidants, procyanidins, epicatechin',
+      doi: '10.1080/07315724.2001.10719181',
+      booktitle:
+        'Ross Products Research Conference on Medical Issues, Synergy in Medical and Nutrition Therapy',
+      year: '2001',
     },
   })
   await prisma.userDocument.create({
@@ -33,59 +111,20 @@ async function seedInternal(prisma: PrismaClient): Promise<void> {
       id: 'ckondtpcn000301mhg9lvaqlu',
       type: 'PhdThesis',
       title:
-        'The postcranial anatomy of \textit{Procolophon} (Parareptilia: Procolophonidae) and its implications for the origin of turtles',
-      author: 'Michael deBraga',
-      citationKey: 'deBraga2001',
-      year: '2001',
+        'Functionality of inulin and polydextrose in stevia or thaumatin sweetened dark chocolate',
+      author: 'Roger Aidoo',
+      year: '2015',
+      abstract:
+        'Chocolate is a high energy product with carbohydrates, including sugar, together with fat, as the main sources of energy. Sucrose is utilized up to 30-60% in chocolate and this confers multiple functional properties on chocolate including sweetness, bulkiness and mouthfeel (texture). Today’s consumers are concerned about the high sugar levels, calories and cariogenicity effects in confectionery products, hence growing the popularity of “light” and “sugar-free” products. This has led to the search for low calorie, low glycemic index, healthier alternatives. Polydextrose and inulin are considered as fibers with many interesting functional attributes that meet the needs of the food industry for healthy foods. Stevia and thaumatin are natural high potency (intense) sweeteners with sensory properties superior to those of other sweeteners. This research investigated the functionality of inulin and polydextrose as sucrose replacers (bulking agents) in sugar-free dark chocolates with stevia or thaumatin as intense sweeteners. The type of bulking agent and concentrations used greatly influenced the rheological properties, textural, melting behaviours and other physical quality characteristics of the developed sugar-free chocolates. Microstructural examination revealed that inulin, which had a lower density than polydextrose tend to have more solids per volume and increased particle volume fraction and solid’s surface area, resulting in a higher particle collision and aggregation thereby limiting chocolate flow. This research also reports for the first time, results of a comparative study of the sweetness brought by the intense sweeteners (stevia and thaumatin) in the sugar-free dark chocolates. With opportunity for improvements in quality of reduced calorie chocolates, findings from this research could be applied to develop or reformulate diabetic and/or reduced calorie chocolates to better meet consumer expectations.',
+      keywords: 'polydextrose, chocolate, inulin, stevia',
       other: {
         create: [
           {
-            field: 'school',
-            value: 'Department of Zoology, University of Toronto',
-          },
-          {
-            field: 'address',
-            value: 'Canada',
+            field: 'institution',
+            value: 'Ghent University',
           },
         ],
       },
-    },
-  })
-  await prisma.userDocument.create({
-    data: {
-      id: 'ckondu6bh000501mh2o2tf00u',
-      type: 'Article',
-      title:
-        'The trouble with login: on usability and computer security in ubiquitous computing',
-      doi: 'http://dx.doi.org/10.1007/s00779-005-0347-6',
-      issn: '1617-4909',
-      number: '6',
-      pages: '357--367',
-      volume: '9',
-      journal: 'Personal Ubiquitous Comput.',
-      priority: 'prio1',
-      year: '2005',
-      other: {
-        create: {
-          field: 'publisher',
-          value: 'Springer-Verlag',
-        },
-      },
-    },
-  })
-  await prisma.userDocument.create({
-    data: {
-      id: 'ckonduhjk000701mh12wia4nf',
-      type: 'Article',
-      author:
-        'Jeff Yan and Alan Blackwell and Ross Anderson and Alasdair Grant',
-      journal: 'IEEE Security and Privacy',
-      year: '2004',
-      volume: '2',
-      pages: '25--31',
-      number: '5',
-      doi: 'http://dx.doi.org/10.1109/MSP.2004.81',
-      issn: '1540-7993',
     },
   })
   await prisma.group.create({
@@ -256,6 +295,8 @@ async function seedInternal(prisma: PrismaClient): Promise<void> {
     'ckondtpcn000301mhg9lvaqlu',
     'ckondu6bh000501mh2o2tf00u',
     'ckonduhjk000701mh12wia4nf',
+    'ckr9eqap6000301mk20hycjqb',
+    'ckr9eq4oc000101mk1ga9bxnt',
   ]) {
     await prisma.userDocument.update({
       where: {
@@ -305,10 +346,14 @@ async function seedInternal(prisma: PrismaClient): Promise<void> {
 }
 
 export async function seed(): Promise<void> {
+  console.log(`Seeding database...`)
   const prisma = new PrismaClient()
 
   try {
     await seedInternal(prisma)
+  } catch (e) {
+    console.error(e)
+    process.exit(1)
   } finally {
     await prisma.$disconnect()
   }
